feat: /places/:placeId/books response data update
Response books info
