+++ src/books/booksController.js
@@ -18,21 +18,6 @@ const booksController = {
             return res.status(500).json(response(baseResponse.SERVER_ERROR));
         }
     },
-
-    getBooksByIsbn: async (req, res) => {
-        try{
-            const isbns = req.query.isbns;
-            if (!isbns)
-                return res.status(404).json(response(baseResponse.NO_BOOK_ISBN));
-            const results = await booksProvider.getBooksByIsbn(isbns.split(','));
-            if (results.error)
-                return res.status(500).json(response(baseResponse.SERVER_ERROR));
-            return res.status(200).json(response(baseResponse.SUCCESS, results));
-        }catch(err){
-            logger.error(err.message);
-            return res.status(500).json(response(baseResponse.SERVER_ERROR));
-        }
-    }
 };
 
 export default booksController;
