+++ src/books/booksRouter.js
@@ -4,6 +4,5 @@ import booksController from './booksController';
 const booksRouter = express.Router();
 
 booksRouter.get('/list', booksController.getBooksByTitle);
-booksRouter.get('/info', booksController.getBooksByIsbn);
 
 export default booksRouter;
